+++ spec/other/clean_spec.rb
@@ -18,20 +18,21 @@ describe "bundle clean" do
   end
 
   it "removes unused gems that are different" do
-    gemfile = <<-G
+    gemfile <<-G
       source "file://#{gem_repo1}"
 
       gem "thin"
       gem "foo"
     G
 
-    install_gemfile(gemfile, :path => "vendor/bundle")
+    bundle "install --path vendor/bundle --no-clean"
 
-    install_gemfile <<-G
+    gemfile <<-G
       source "file://#{gem_repo1}"
 
       gem "thin"
     G
+    bundle "install --no-clean"
 
     bundle :clean
 
@@ -44,21 +45,22 @@ describe "bundle clean" do
   end
 
   it "removes old version of gem if unused" do
-    gemfile = <<-G
+    gemfile <<-G
       source "file://#{gem_repo1}"
 
       gem "rack", "0.9.1"
       gem "foo"
     G
 
-    install_gemfile(gemfile, :path => "vendor/bundle")
+    bundle "install --path vendor/bundle --no-clean"
 
-    install_gemfile <<-G
+    gemfile <<-G
       source "file://#{gem_repo1}"
 
       gem "rack", "1.0.0"
       gem "foo"
     G
+    bundle "install --no-clean"
 
     bundle :clean
 
@@ -71,21 +73,22 @@ describe "bundle clean" do
   end
 
   it "removes new version of gem if unused" do
-    gemfile = <<-G
+    gemfile <<-G
       source "file://#{gem_repo1}"
 
       gem "rack", "1.0.0"
       gem "foo"
     G
 
-    install_gemfile(gemfile, :path => "vendor/bundle")
+    bundle "install --path vendor/bundle --no-clean"
 
-    install_gemfile <<-G
+    gemfile <<-G
       source "file://#{gem_repo1}"
 
       gem "rack", "0.9.1"
       gem "foo"
     G
+    bundle "install --no-clean"
 
     bundle :clean
 
@@ -98,7 +101,7 @@ describe "bundle clean" do
   end
 
   it "remove gems in bundle without groups" do
-    gemfile = <<-G
+    gemfile <<-G
       source "file://#{gem_repo1}"
 
       gem "foo"
@@ -108,8 +111,8 @@ describe "bundle clean" do
       end
     G
 
-    install_gemfile(gemfile, :path => "vendor/bundle")
-    bundle "install --without test_group"
+    bundle "install --path vendor/bundle --no-clean"
+    bundle "install --without test_group --no-clean"
     bundle :clean
 
     out.should be == "Removing rack (1.0.0)"
@@ -125,7 +128,7 @@ describe "bundle clean" do
     revision = revision_for(lib_path("foo-1.0"))
     git_path = lib_path('foo-1.0')
 
-    gemfile = <<-G
+    gemfile <<-G
       source "file://#{gem_repo1}"
 
       gem "rack", "1.0.0"
@@ -134,7 +137,7 @@ describe "bundle clean" do
       end
     G
 
-    install_gemfile(gemfile, :path => "vendor/bundle")
+    bundle "install --path vendor/bundle --no-clean"
 
     bundle :clean
 
@@ -147,7 +150,7 @@ describe "bundle clean" do
     revision = revision_for(lib_path("foo-1.0"))
     git_path = lib_path('foo-1.0')
 
-    gemfile = <<-G
+    gemfile <<-G
       source "file://#{gem_repo1}"
 
       gem "rack", "1.0.0"
@@ -156,13 +159,14 @@ describe "bundle clean" do
       end
     G
 
-    install_gemfile(gemfile, :path => "vendor/bundle")
+    bundle "install --path vendor/bundle --no-clean"
 
-    install_gemfile <<-G
+    gemfile <<-G
       source "file://#{gem_repo1}"
 
       gem "rack", "1.0.0"
     G
+    bundle "install --no-clean"
 
     bundle :clean
 
@@ -182,7 +186,7 @@ describe "bundle clean" do
     build_git "foo"
     revision = revision_for(lib_path("foo-1.0"))
 
-    gemfile = <<-G
+    gemfile <<-G
       source "file://#{gem_repo1}"
 
       gem "rack", "1.0.0"
@@ -191,13 +195,12 @@ describe "bundle clean" do
       end
     G
 
-    install_gemfile(gemfile, :path => "vendor/bundle")
+    bundle "install --path vendor/bundle --no-clean"
 
     update_git "foo"
     revision2 = revision_for(lib_path("foo-1.0"))
 
-    bundle :update
-    bundle :install
+    bundle "update --no-clean"
     bundle :clean
 
     out.should be == "Removing foo (1.0 #{revision[0..11]})"
@@ -222,7 +225,7 @@ describe "bundle clean" do
       gem "activesupport", :git => "#{lib_path('rails')}", :ref => '#{revision}'
     G
 
-    bundle "install --path vendor/bundle"
+    bundle "install --path vendor/bundle --no-clean"
     bundle :clean
     out.should be == ""
 
@@ -243,20 +246,21 @@ describe "bundle clean" do
 
   # handling bundle clean upgrade path from the pre's
   it "removes .gem/.gemspec file even if there's no corresponding gem dir is already moved" do
-    gemfile = <<-G
+    gemfile <<-G
       source "file://#{gem_repo1}"
 
       gem "thin"
       gem "foo"
     G
 
-    install_gemfile(gemfile, :path => "vendor/bundle")
+    bundle "install --path vendor/bundle --no-clean"
 
-    install_gemfile <<-G
+    gemfile <<-G
       source "file://#{gem_repo1}"
 
       gem "foo"
     G
+    bundle "install --no-clean"
 
     FileUtils.rm(vendored_gems("bin/rackup"))
     FileUtils.rm_rf(vendored_gems("gems/thin-1.0"))
@@ -268,5 +272,103 @@ describe "bundle clean" do
     should_have_gems 'foo-1.0'
 
     vendored_gems("bin/rackup").should_not exist
+  end
+
+  it "does not call clean automatically when using system gems" do
+    gemfile <<-G
+      source "file://#{gem_repo1}"
+
+      gem "thin"
+      gem "rack"
+    G
+    bundle :install
+
+    gemfile <<-G
+      source "file://#{gem_repo1}"
+
+      gem "rack"
+    G
+    bundle :install
+
+    sys_exec "gem list"
+    out.should include("rack (1.0.0)")
+    out.should include("thin (1.0)")
+  end
+
+  it "--clean should override the bundle setting" do
+    gemfile <<-G
+      source "file://#{gem_repo1}"
+
+      gem "thin"
+      gem "rack"
+    G
+    bundle "install --path vendor/bundle --no-clean"
+
+    gemfile <<-G
+      source "file://#{gem_repo1}"
+
+      gem "rack"
+    G
+    bundle "install --clean"
+
+    should_have_gems 'rack-1.0.0'
+    should_not_have_gems 'thin-1.0'
+  end
+
+  it "clean automatically on --path" do
+    gemfile <<-G
+      source "file://#{gem_repo1}"
+
+      gem "thin"
+      gem "rack"
+    G
+    bundle "install --path vendor/bundle"
+
+    gemfile <<-G
+      source "file://#{gem_repo1}"
+
+      gem "rack"
+    G
+    bundle "install"
+
+    should_have_gems 'rack-1.0.0'
+    should_not_have_gems 'thin-1.0'
+  end
+
+  it "cleans on bundle update with --path" do
+    build_repo2
+
+    gemfile <<-G
+      source "file://#{gem_repo2}"
+
+      gem "foo"
+    G
+    bundle "install --path vendor/bundle"
+
+    update_repo2 do
+      build_gem 'foo', '1.0.1'
+    end
+
+    bundle :update
+    should_not_have_gems 'foo-1.0'
+  end
+
+  it "does not clean on bundle update when using --system" do
+    build_repo2
+
+    gemfile <<-G
+      source "file://#{gem_repo2}"
+
+      gem "foo"
+    G
+    bundle "install"
+
+    update_repo2 do
+      build_gem 'foo', '1.0.1'
+    end
+    bundle :update
+
+    sys_exec "gem list"
+    out.should include("foo (1.0.1, 1.0)")
   end
 end